リダイレクト処理のファイル分割
