+++ microshell/src/process.rs
@@ -3,3 +3,4 @@ pub mod builtin_cmd;
 pub mod display;
 pub mod parse;
 pub mod fork;
+pub mod redirect;
